backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
mysql-connector-python==8.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
requests==2.31.0

# Computer Vision Pipeline Dependencies
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.24.3
mediapipe==0.10.9

